lambda/requirements.txt
boto3==1.34.144
cassandra_driver==3.29.1
langchain==0.2.13
langchain_community==0.2.12
langchain_core==0.2.30
opensearch_py==2.6.0
cassio==0.1.8